Stop forwarding options to remark/rehype plugins (#576)
This no longer serves its initial purpose, and is definitely unexpected
behavior that can break plugins which expect certain properties or,
worse, if a MDX option happen to match an option of that plugin.

This change shouldn't introduce problems because it's hard to imagine
why someone would rely on this undocumented and buggy behavior instead
of filing an issue in this very actively maintained project. Also, if
someone was getting around this bug by always forcing custom options,
that workaround will continue to work after this change.